requirements.txt
psycopy2-binary
hdfs==2.5.8
cassandra-driver
pyhive